General clean up and overhaul - use dynamic document for references - move test disk operations to store module - make test store operations use vcs
